pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crack-synth"
version = "0.1.0"
description = "Task-oriented crack synthesis workspace with a diffusion baseline method."
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
  "accelerate>=0.30",
  "diffusers>=0.29",
  "numpy>=1.24",
  "opencv-python-headless>=4.8",
  "pandas>=2.0",
  "Pillow>=10.0",
  "PyYAML>=6.0",
  "safetensors>=0.4",
  "torchvision>=0.16",
  "tqdm>=4.66",
  "transformers>=4.41",
]

[project.scripts]
build_manifests = "crack_synth.methods.diffusion_baseline.build_manifests:main"
prepare_rois = "crack_synth.methods.diffusion_baseline.prepare_rois:main"
generate_baseline = "crack_synth.methods.diffusion_baseline.generate_baseline:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
